requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
apispec==6.8.1
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
mando==0.7.1
MarkupSafe==3.0.2
marshmallow==3.25.1
multidict==6.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
propcache==0.2.1
pycparser==2.22
PyMySQL==1.1.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.37
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
webargs==8.6.0
Werkzeug==3.1.3
yarl==1.18.3
